bot/requirements.txt
# Bot Dependencies for Open-LLM-VTuber
# Required packages for Telegram and Discord bots

# Core dependencies
websockets>=11.0.0
numpy>=1.21.0
pillow>=9.0.0
pydub>=0.25.0
loguru>=0.6.0
aiohttp>=3.8.0

# Telegram bot dependencies
python-telegram-bot>=20.0

# Discord bot dependencies
discord.py>=2.3.0

# Audio processing
ffmpeg-python>=0.2.0
